+++ x/stakeibc/client/cli/tx.go
@@ -598,3 +598,32 @@ func CmdUpdateInnerRedemptionRateBounds() *cobra.Command {
 
 	return cmd
 }
+
+func CmdResumeHostZone() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "resume-host-zone [chainid]",
+		Short: "Broadcast message resume-host-zone",
+		Args:  cobra.ExactArgs(1),
+		RunE: func(cmd *cobra.Command, args []string) (err error) {
+			argChainId := args[0]
+
+			clientCtx, err := client.GetClientTxContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			msg := types.NewMsgResumeHostZone(
+				clientCtx.GetFromAddress().String(),
+				argChainId,
+			)
+			if err := msg.ValidateBasic(); err != nil {
+				return err
+			}
+			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
+		},
+	}
+
+	flags.AddTxFlagsToCmd(cmd)
+
+	return cmd
+}
